Implement working seatbelt sandbox
Implement a working sandbox using macOS seatbelt sandboxing. Use parameters to
communicate paths to the sandbox. Intercept calls chngproc and fileproc for
that purpose.

Note: We're adding realpath to AC_CHECK_FUNCS in configure.ac even
though we only use it on macOS with seatbelt sandboxing. Linux has it
with the same semantics relevant to our usage so we don't bother
conditionalising that for now.