Update upstream source
Upstream has fixed some type warnings and is using the OpenSSL 1.1 API
themselves now, as LibreSSL recently gained support for it. This makes
two of our local patches redundant. It likely also raises our minimum
version dependencies, at the very least to 1.1 for OpenSSL. We'll dig
into that if and when that becomes a problem.